refactor: remove ui composableHelpers circular dependency
